requirements.txt
click==6.7
gordon-dns==0.0.1.dev8
setuptools==38.2.5
toml==0.9.3.1
ulogger==1.0.1
zope.interface==4.4.3
